server/grpc_service/requirements.txt
grpcio
grpcio-tools
protobuf
python-dotenv
psycopg2-binary
